shared/idmap: Export non-Linux specific logic
